transfers_rework: Copy Transfers endpoint + tests

+++ src/transfers/views/transfer_viewset.py
@@ -43,7 +43,7 @@ class TransferViewSet(ModelViewSet):
                 "objects_ids": openapi.Schema(
                     type=openapi.TYPE_ARRAY,
                     items=openapi.Items(type=openapi.TYPE_INTEGER),
-                    description="List of IDs to delete",
+                    description="List of IDs to delete.",
                 )
             },
             required=["objects_ids"],
@@ -52,14 +52,58 @@ class TransferViewSet(ModelViewSet):
     @action(detail=False, methods=["delete"])
     def bulk_delete(self, request, budget_pk: str) -> Response:
         """
-        Removes multiple Transfer with given IDs at once.
+        Removes multiple Transfers with given IDs at once.
 
         Returns:
-            Response: Filtered TransferCategory QuerySet.
+            Response: API response with status.
         """
         ids = request.data.get("objects_ids", None)
         if not isinstance(ids, list):
             return Response({"error": "objects_ids must be a list."}, status=status.HTTP_400_BAD_REQUEST)
+        if not ids:
+            return Response({"error": "objects_ids must not be an empty list."}, status=status.HTTP_400_BAD_REQUEST)
         with transaction.atomic():
             self.serializer_class.Meta.model.objects.filter(period__budget__id=int(budget_pk), id__in=ids).delete()
-        return Response({}, status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+    @swagger_auto_schema(
+        method="post",
+        request_body=openapi.Schema(
+            type=openapi.TYPE_OBJECT,
+            properties={
+                "objects_ids": openapi.Schema(
+                    type=openapi.TYPE_ARRAY,
+                    items=openapi.Items(type=openapi.TYPE_INTEGER),
+                    description="List of IDs to copy.",
+                )
+            },
+            required=["objects_ids"],
+        ),
+    )
+    @action(detail=False, methods=["post"])
+    def copy(self, request, budget_pk: str) -> Response:
+        """
+        Copies multiple Transfers with given IDs.
+
+        Returns:
+            Response: API response with status.
+        """
+        ids = request.data.get("objects_ids", None)
+        if not isinstance(ids, list):
+            return Response({"error": "objects_ids must be a list."}, status=status.HTTP_400_BAD_REQUEST)
+        if not ids:
+            return Response({"error": "objects_ids must not be an empty list."}, status=status.HTTP_400_BAD_REQUEST)
+        new_objects = [
+            self.serializer_class.Meta.model(
+                **{
+                    field_name: getattr(copied_object, field_name)
+                    for field_name in self.serializer_class.Meta.fields
+                    if field_name not in ("id",)
+                }
+            )
+            for copied_object in self.serializer_class.Meta.model.objects.filter(
+                period__budget__id=int(budget_pk), id__in=ids
+            )
+        ]
+        objs = self.serializer_class.Meta.model.objects.bulk_create(new_objects)
+        return Response({"ids": [obj.id for obj in objs]} if objs else [], status=status.HTTP_201_CREATED)
